apps/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.1
sqlalchemy==2.0.31
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.4
pika==1.3.2
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
cryptography==42.0.8
python-dotenv==1.0.1
prometheus-client==0.20.0
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
testcontainers[postgres,redis]==3.7.1
